requirements.txt
httpx
keyring
keyrings.alt
beautifulsoup4
pandas
pycryptodomex
pydantic
questionary
rich
Flask